JSONToken should always contain scopes.
This is already the case when writing a token file. Presumably the only
reason it was an Option was for backwards compatibility, but we're
already breaking compatibility with the change to the hash value so this
seems like an appropriate time to make the change.

This change also highlights how unused the hash value has been
previously. Future changes plan to use the hash value for more efficient
handling.